requirements.txt
streamlit==1.31.1
openai==1.12.0
python-dotenv==1.0.1
reportlab==4.1.0
python-slugify==8.0.4
markdown2==2.4.12 